Mv InstallInfo to models.models
