+++ tests/ansible_galaxy/models/test_install_info_model.py
@@ -4,31 +4,31 @@ import logging
 import attr
 import pytest
 
-from ansible_galaxy.models import install_info
+from ansible_galaxy.models.models import InstallInfo
 
 log = logging.getLogger(__name__)
 
 
 def test_init():
     install_date = datetime.datetime.utcnow()
-    info = install_info.InstallInfo(version='1.2.3',
-                                    install_date=install_date.strftime('%c'),
-                                    install_date_iso=install_date)
+    info = InstallInfo(version='1.2.3',
+                       install_date=install_date.strftime('%c'),
+                       install_date_iso=install_date)
 
     log.debug('info: %s', info)
 
-    assert isinstance(info, install_info.InstallInfo)
+    assert isinstance(info, InstallInfo)
     assert isinstance(attr.asdict(info), dict)
 
 
 def test_from_version_date():
     install_datetime = datetime.datetime.utcnow()
-    info = install_info.InstallInfo.from_version_date(version='1.2.3',
-                                                      install_datetime=install_datetime)
+    info = InstallInfo.from_version_date(version='1.2.3',
+                                         install_datetime=install_datetime)
 
     log.debug('info: %s', info)
 
-    assert isinstance(info, install_info.InstallInfo)
+    assert isinstance(info, InstallInfo)
     assert isinstance(attr.asdict(info), dict)
 
     assert info.install_date_iso == install_datetime
@@ -36,9 +36,9 @@ def test_from_version_date():
 
 def test_frozen():
     install_date = datetime.datetime.utcnow()
-    info = install_info.InstallInfo(version='1.2.3',
-                                    install_date=install_date.strftime('%c'),
-                                    install_date_iso=install_date)
+    info = InstallInfo(version='1.2.3',
+                       install_date=install_date.strftime('%c'),
+                       install_date_iso=install_date)
 
     log.debug('info: %s', info)
 
@@ -54,17 +54,17 @@ def test_frozen():
 
 def test_equal():
     install_date1 = datetime.datetime.utcnow()
-    info1 = install_info.InstallInfo(version='1.2.3',
-                                     install_date=install_date1.strftime('%c'),
-                                     install_date_iso=install_date1)
+    info1 = InstallInfo(version='1.2.3',
+                        install_date=install_date1.strftime('%c'),
+                        install_date_iso=install_date1)
 
-    info1a = install_info.InstallInfo(version='1.2.3',
-                                      install_date=install_date1.strftime('%c'),
-                                      install_date_iso=install_date1)
+    info1a = InstallInfo(version='1.2.3',
+                         install_date=install_date1.strftime('%c'),
+                         install_date_iso=install_date1)
 
-    info2 = install_info.InstallInfo(version='9.9.9',
-                                     install_date=install_date1.strftime('%c'),
-                                     install_date_iso=install_date1)
+    info2 = InstallInfo(version='9.9.9',
+                        install_date=install_date1.strftime('%c'),
+                        install_date_iso=install_date1)
 
     assert info1 == info1a
     assert info1a == info1
